% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_and_clean.R
\name{fetch_and_clean}
\alias{fetch_and_clean}
\title{Fetch and clean search results}
\usage{
fetch_and_clean(
  keyword,
  field,
  db,
  sdkey,
  no_duplicate,
  limit_per_search = NULL,
  start_year = NULL,
  end_year = NULL,
  additional_args = list()
)
}
\arguments{
\item{keyword}{Required. A string value.}

\item{field}{The field to search for the keywords. Currently support:
\code{abstract}, \code{all}
Some databases may not support some field to be used in the search.}

\item{db}{A string of one database name. Default to \code{sage_journal}.
Currently support: \code{sage_journal}, \code{science_direct}, \code{pubmed}, \code{proquest}
If both correct and wrong database names are specified, only the correct ones
will be queried.}

\item{sdkey}{API key for science direct.}

\item{no_duplicate}{Whether to remove duplicated results (TRUE) or not (FALSE).
One article could be returned by multiple databases, if you specified multiple
databases to search in \code{database_name}.}

\item{limit_per_search}{The maximum number of results collected from each
database.}
}
\description{
Fetches and cleans search results from PubMed as a dataframe. This is used by
the main web_scrapper() function.
If 0 search result, return an empty dataframe.
}
