% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_proquest_subdb}
\alias{get_proquest_subdb}
\title{Get Proquest Sub-Daatabases}
\usage{
get_proquest_subdb()
}
\description{
This function retrieves all the subscribed Proquest Sub-Databases by sending a query request to
a non-existing sub-db and parsing the returned message where all subscriptions are mentioned.
}
