% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_database.R
\name{search_database}
\alias{search_database}
\title{Search Specified Databases}
\usage{
search_database(
  keywords,
  relationship = "or",
  field = "abstract",
  database_name = c("sage_journal", "science_direct", "pubmed", "proquest"),
  sdkey = "",
  start_year = NULL,
  end_year = NULL,
  additional_args = list(),
  limit_per_search = NULL,
  drop_duplicate = T,
  availability_pubmed = T
)
}
\arguments{
\item{keywords}{Required.
If you only have one set of keywords, this will be a string of one keyword or
a vector of multiple keywords:
"machine learning"
c("neural networks", "deep learning")
If you have multiple sets of keywords, for example each set describing one concept
using different wording, and you need to search all the combinations between these
keyword sets, then provide a list of keywords:
list(c("neural networks", "deep learning"),
c("text","textual","document"),
c("psychology","psychological"))
Note that in the second case, argument "relationship" is ignored as it is useless.}

\item{relationship}{Relationship between the keywords. Can be either \code{or} or
\code{and}. Default to \code{or}. When using \code{or}, the function runs a search for each
and every keyword independently. When using \code{and}, the function runs one search
that combines all the keywords.}

\item{field}{The field to search for the keywords. Currently support:
\code{abstract}, \code{all}
Some databases may not support some field to be used in the search.}

\item{database_name}{A string of one database name or a vector of multiple
database names. Default to \code{sage_journal}. Currently support:
\code{sage_journal}, \code{science_direct}, \code{pubmed}, \code{proquest}
If both correct and wrong database names are specified, only the correct ones
will be queried.}

\item{sdkey}{API key for science direct.}

\item{limit_per_search}{The maximum number of results collected from each
database.}

\item{drop_duplicate}{Whether to remove duplicated results (TRUE) or not (FALSE).
One article could be returned by multiple databases, if you specified multiple
databases to search in \code{database_name}.}

\item{availability_pubmed}{Whether query to see if the full text of an article
is available to you (TRUE) or not (FALSE). Default to FALSE. Because it has
to check each article one by one, it can be time consuming and can hit the
3 queries / sec limit.}
}
\description{
This is the main function to be used. This function searches the specified
databases using the keywords provided.
}
